requirements.txt
argon2-cffi
pycryptodome
PyNaCl